Move tests over to remote settings folder.
