+++ experimenter/tests/integration/nimbus/test_experimenter_ui.py
@@ -1,5 +1,4 @@
 import os
-from urllib.parse import urljoin
 
 import pytest
 
@@ -107,60 +106,3 @@ def test_every_form_page_can_be_resaved(
     audience = metrics.save_and_continue()
     summary = audience.save_and_continue()
     assert summary.experiment_slug is not None
-
-
-@pytest.mark.nimbus_ui
-def test_rollout_create_and_update(
-    selenium,
-    base_url,
-    create_experiment,
-    kinto_client,
-    experiment_name,
-    slugify,
-):
-    experiment_slug = str(slugify(experiment_name))
-    create_experiment(selenium, is_rollout=True).launch_and_approve()
-
-    kinto_client.approve()
-    summary = SummaryPage(selenium, urljoin(base_url, experiment_slug)).open()
-
-    summary.wait_for_live_status()
-    audience = summary.navigate_to_audience()
-
-    audience.percentage = "60"
-    audience.save_and_continue()
-
-    summary_page = SummaryPage(selenium, urljoin(base_url, experiment_slug)).open()
-    summary_page.wait_for_update_request_visible()
-
-
-@pytest.mark.nimbus_ui
-def test_rollout_live_update_reject_on_experimenter(
-    selenium,
-    base_url,
-    create_experiment,
-    kinto_client,
-    experiment_name,
-    slugify,
-):
-    experiment_slug = str(slugify(experiment_name))
-    create_experiment(selenium, is_rollout=True).launch_and_approve()
-
-    kinto_client.approve()
-    summary = SummaryPage(selenium, urljoin(base_url, experiment_slug)).open()
-
-    summary.wait_for_live_status()
-    audience = summary.navigate_to_audience()
-
-    audience.percentage = "60"
-    audience.save_and_continue()
-
-    summary_page = SummaryPage(selenium, urljoin(base_url, experiment_slug)).open()
-    summary_page.wait_for_update_request_visible()
-
-    summary_page.request_update_and_reject()
-    summary_page.wait_for_rejection_reason_text_input_visible()
-
-    summary_page.set_rejection_reason()
-    summary_page.submit_rejection()
-    summary_page.wait_for_rejection_notice_visible()
